[Docs] Fix a few out-of-date docs
Closes #1179. Plus sundry other docs fixes.

`kEntityAccessError` still referred to `Context.access`, which was
removed in #1054.

The exception raised, when `register()` is called with mismatched
entity reference and traits data list lengths, was changed from
`std::out_of_range` to our custom `InputValidationException` in #1071.

Windows `FileUrlPathConverter.pathFromUrl()` handling for `file://` URLs
that, when decoded to a file path, exceeds the Windows MAX_PATH limit,
was changed in #1257 to automatically convert to UNC device path style
`\\?\`.

The `Regex` PCRE2 wrapper class, when initially written, was stateful
and so not thread-safe. However, that was changed upon review
(https://github.com/OpenAssetIO/OpenAssetIO/pull/1246#discussion_r1465151326),
but the docstring was not updated to reflect that.

Signed-off-by: David Feltell <david.feltell@foundry.com>