[Docs] Update ROADMAP
The roadmap was woefully out of date.

So restructure to make slightly more maintainable and conform to ASWF
Incubation guidelines, i.e.

> What use cases are possible now?
> What does the next year look like in terms of additional features and
> use cases covered?

Link to GitHub milestones rather than listing features, for easier
maintainability of future use-cases. Similarly link to
`examples/README.md` for easier maintainability of current use-cases.

Add some generic blub summarising the state of the API, structured with
the goal of not having to update it often.

For future plans, reference the GitHub milestones, but also highlight
the drive to adoption.

Signed-off-by: David Feltell <david.feltell@foundry.com>